Collect all embedding code into a single module
