+++ autogpt/memory/local.py
@@ -7,8 +7,8 @@ from typing import Any, List
 import numpy as np
 import orjson
 
-from autogpt.llm_utils import create_embedding_with_ada
 from autogpt.memory.base import MemoryProviderSingleton
+from autogpt.memory.embeddings import get_ada_embedding
 
 EMBED_DIM = 1536
 SAVE_OPTIONS = orjson.OPT_SERIALIZE_NUMPY | orjson.OPT_SERIALIZE_DATACLASS
