+++ autogpt/memory/redismem.py
@@ -10,9 +10,9 @@ from redis.commands.search.field import TextField, VectorField
 from redis.commands.search.indexDefinition import IndexDefinition, IndexType
 from redis.commands.search.query import Query
 
-from autogpt.llm_utils import create_embedding_with_ada
 from autogpt.logs import logger
 from autogpt.memory.base import MemoryProviderSingleton
+from autogpt.memory.embeddings import create_embedding_with_ada
 
 SCHEMA = [
     TextField("data"),
