+++ autogpt/memory/weaviate.py
@@ -1,12 +1,10 @@
-import uuid
-
 import weaviate
 from weaviate import Client
 from weaviate.embedded import EmbeddedOptions
 from weaviate.util import generate_uuid5
 
-from autogpt.config import Config
-from autogpt.memory.base import MemoryProviderSingleton, get_ada_embedding
+from autogpt.memory.base import MemoryProviderSingleton
+from autogpt.memory.embeddings import get_ada_embedding
 
 
 def default_schema(weaviate_index):
